Add files via upload
Updated EndGame to add a better end game menu. Now it compares your score based on your username.

Additionally, made it so that if no username was entered, or the Username "Username" was entered, it would keep you as a guest account and not save your score.

Refactored Cell by adding a new Class, Colours, which contains a hashmap for all the colours of the cells.